requirements.txt
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pdoc==12.3.1
Pygments==2.14.0
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.2.2
